# documentation/api/openapi.yaml
openapi: "3.0.0"
info:
  version: "1.0.1"
  title: Pivchart API
  description: Multi-file boilerplate for OpenAPI Specification.
  license:
    name: MIT
  contact:
    name: API Support
    url: http://www.example.com/support
    email: support@example.com
servers:
  - url: http://api.pivchart.com
...
paths:
...
  /drinks/{drinkId}:
    get:
      tags:
        - Drinks
      summary: "Returns drink with given id"
      parameters:
        - $ref: "#/components/parameters/DrinkId"

      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrinkResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceNotFoundResponse"
    put:
      tags:
        - Drinks
      summary: "Updates drink with given id"
      parameters:
        - $ref: "#/components/parameters/DrinkId"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDrinkRequest'

      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrinkResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceNotFoundResponse"
    delete:
      tags:
        - Drinks
      summary: "Deletes drink by id"
      parameters:
        - $ref: "#/components/parameters/DrinkId"
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceNotFoundResponse"
...
  /gatherings/{gatheringId}:
    get:
      tags:
        - Gatherings
      summary: "Returns gathering with given id"
      parameters:
        - $ref: "#/components/parameters/GatheringId"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatheringResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceNotFoundResponse"
    put:
      tags:
        - Gatherings
      summary: "Updates gathering with given id"

      parameters:
        - $ref: "#/components/parameters/GatheringId"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGatheringRequest'

      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatheringResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceNotFoundResponse"
    delete:
      tags:
        - Gatherings
      summary: "Delete gathering by id"
      parameters:
        - $ref: "#/components/parameters/GatheringId"

      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceNotFoundResponse"
  /gatherings/{gatheringId}/finish:
    post:
      tags:
        - Gatherings
      summary: "Finish gathering by id"

      parameters:
        - $ref: "#/components/parameters/GatheringId"
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceNotFoundResponse"
...
  /gatherings/{gatheringId}/drinks/{gatheringDrinkId}:
    delete:
      tags:
        - Gathering drinks
      summary: "Deletes drink from gathering"
      parameters:
        - $ref: "#/components/parameters/GatheringId"
        - $ref: "#/components/parameters/GatheringDrinkId"

      responses:
        '204':
          description: No Content
        '400':
          description: Bad request - drink may not be from this gahering
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceNotFoundResponse"
...
  /gatherings/{gatheringId}/members/{gatheringMemberId}:
    delete:
      tags:
        - Gathering members
      summary: "Delete member from gathering"
      parameters:
        - $ref: "#/components/parameters/GatheringId"
        - $ref: "#/components/parameters/GatheringMemberId"

      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceNotFoundResponse"
...
  /competitions/{competitionId}/drinks/{competitionDrinkId}:
    delete:
      tags:
        - Competition Drinks
      summary: "Delete drink from competition"
      parameters:
        - $ref: "#/components/parameters/CompetitionId"
        - $ref: "#/components/parameters/CompetitionDrinkId"

      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'

  #  /competitions/{competitionId}/members:
  #    get:
  #      tags:
  #        - competitions-member-controller
  #      summary: "Get all members in competition"
  #      parameters:
  #        -  in: path
  #           name: competitionId
  #           description: competitionId
  #           required: true
  #           schema:
  #             type: integer
  #
  #      responses:
  #        '200':
  #          description: OK
  #          content:
  #            application/json:
  #              schema:
  #                type: array
  #                items:
  #                  $ref: '#/components/schemas/CompetitionMemberResponse'
  #        '401':
  #          description: Unauthorized
  #        '403':
  #          description: Forbidden
  #        '404':
  #          description: Not Found
  #    post:
  #      tags:
  #        - competitions-member-controller
  #      summary: "Add member to competition"
  #
  #      parameters:
  #        - in: path
  #          name: competitionId
  #          description: competitionId
  #          required: true
  #          schema:
  #            type: integer
  #
  #      requestBody:
  #        required: true
  #        content:
  #          application/json:
  #            schema:
  #              $ref: '#/components/schemas/CreateCompetitionMemberRequest'
  #
  #      responses:
  #        '200':
  #          description: OK
  #          content:
  #            application/json:
  #              schema:
  #                $ref: '#/components/schemas/CompetitionMemberResponse'
  #        '201':
  #          description: Created
  #        '401':
  #          description: Unauthorized
  #        '403':
  #          description: Forbidden
  #        '404':
  #          description: Not Found
  #
  #  /competitions/{competitionId}/members/{competitionMemberId}:
  #    delete:
  #      tags:
  #        - competitions-member-controller
  #      summary: "Delete member from competition"
  #      parameters:
  #        - in: path
  #          name: competitionId
  #          description: competitionId
  #          required: true
  #          schema:
  #            type: integer
  #
  #        - name: competitionMemberId
  #          in: path
  #          description: competitionMemberId
  #          required: true
  #          schema:
  #            type: integer
  #
  #      responses:
  #        '204':
  #          description: No Content
  #        '401':
  #          description: Unauthorized
  #        '403':
  #          description: Forbidden
...
  /pubs/{pubId}/drinks/{pubDrinkId}:
    get:
      tags:
        - Pub Drinks
      summary: Returns pub drink with given id
      parameters:
        - $ref: "#/components/parameters/PubId"
        - $ref: "#/components/parameters/PubDrinkId"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PubDrinkResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'
    put:
      tags:
        - Pub Drinks
      summary: Updates pub drink with given id
      parameters:
        - $ref: "#/components/parameters/PubId"
        - $ref: "#/components/parameters/PubDrinkId"

      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePubDrinkRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PubDrinkResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'
    delete:
      tags:
        - Pub Drinks
      summary: Deletes pub drink with given id
      parameters:
        - $ref: "#/components/parameters/PubId"
        - $ref: "#/components/parameters/PubDrinkId"
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundResponse'
...
components:
  parameters:
    DrinkId:
      in: path
      name: drinkId
      description: drinkId
      required: true
      schema:
        type: integer
    GatheringId:
      name: gatheringId
      in: path
      description: gatheringId
      required: true
      schema:
        type: integer
    GatheringDrinkId:
      name: gatheringDrinkId
      in: path
      description: gatheringDrinkId
      required: true
      schema:
        type: integer
    GatheringMemberId:
      name: gatheringMemberId
      in: path
      description: gatheringMemberId
      required: true
      schema:
        type: integer
    CompetitionId:
      name: competitionId
      in: path
      description: competitionId
      required: true
      schema:
        type: integer
    CompetitionDrinkId:
      in: path
      name: competitionDrinkId
      description: competitionDrinkId
      required: true
      schema:
        type: integer
...
    PubId:
      in: path
      name: pubId
      description: pubId
      required: true
      schema:
        type: integer
    PubDrinkId:
      in: path
      name: pubDrinkId
      description: pubDrinkId
      required: true
      schema:
        type: integer
  schemas:
...
    DrinkResponse:
      type: object
      properties:
        id:
          type: integer
          example: 5
        name:
          type: string
          example: Teleport
        alcoholVolume:
          type: number
          format: double
          example: 60.0

    # Gathering
...
    GatheringResponse:
      type: object
      properties:
        id:
          type: integer
          example: 42
        title:
          type: string
          example: Gathering@Five
        finished:
          type: boolean
        public:
          type: boolean
        pub:
          $ref: '#/components/schemas/PubResponse'
        startDateTime:
          $ref: "#/components/schemas/DateTime"

      # Competition
...
    CompetitionMemberResponse:
      type: object
      properties:
        competition:
          $ref: '#/components/schemas/CompetitionResponse'
        owner:
          type: boolean
        user:
          $ref: '#/components/schemas/UserResponse'
...
    CompetitionResponse:
      type: object
      properties:
        id:
          type: integer
          example: 187
        title:
          type: string
        startDateTime:
          $ref: "#/components/schemas/DateTime"
        endDateTime:
          $ref: "#/components/schemas/DateTime"
        endType:
          $ref: "#/components/schemas/CompetitionEndType"
        finished:
          type: boolean
        pub:
          $ref: '#/components/schemas/PubResponse'
    PubDrinkResponse:
      type: object
      properties:
        id:
          type: integer
          example: 147
        pubId:
          type: number
          example: 55
        price:
          type: number
          format: double
          example: 155
        drink:
          $ref: '#/components/schemas/DrinkResponse'
    PubResponse:
      type: object
      properties:
        id:
          type: integer
          example: 13
        name:
          type: string
          example: "Propaganda"
        country:
          type: string
          example: "Czech republic"
        city:
          type: string
          example: "Prague"
        street:
          type: string
          example: "Pštrossova 220/29"
        postalCode:
          type: string
          example: "11000"
        latitude:
          type: number
          format: double
          example: 50.079995087080945
        longitude:
          type: number
          format: double
          example: 14.415304373330383

      # Drink
    UpdateDrinkRequest:
      type: object
      properties:
        name:
          type: string
        alcoholVolume:
          type: number
          format: double

      # Gathering
    UpdateGatheringRequest:
      required:
        - title
      type: object
      properties:
        title:
          type: string
          example: "Gathering@Five"
        public:
          type: boolean

      # Competition
...
    UpdatePubDrinkRequest:
      type: object
      properties:
        price:
          type: integer
          example: 155
        volume:
          type: integer
          example: 20
...
    DateTime:
      type: string
      example: 2022-04-16T14:14:17
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 75
        username:
          type: string
          example: "hugo.mrazek"
        email:
          type: string
          example: "hugo.mrazek@gmail.com"
        sex:
          type: string
          example: "MALE"
        weight:
          type: number
          format: double
          example: 97.5
    CompetitionEndType:
      type: string
      enum: [ "LEADER_STOPS", "TIME_ATTACK", "AMOUNT" ]
    ResourceNotFoundResponse:
      type: object
      properties:
        message:
          type: string
          example: "Resource not found"
...
    ForbiddenResponse:
      type: object
      properties:
        message:
          type: string
          example: "Can't touch this"
    UnauthorizedResponse:
      type: object
      properties:
        message:
          type: string
          example: "Wait, that's illegal"
    BadRequestResponse:
      type: object
      properties:
        message:
          type: string
          example: "Bad request"
        errors:
          type: object
          description: Note that response differs for each request based on the required values of given request.
          example:
            name: "Name is required"
            alcoholVolume: "Alcohol volume needs to be a positive integer"